Rename loaders.ts to createStaticLoader.ts
