+++ src/storybook/utils/loaders/index.ts
@@ -1,1 +1,1 @@
-export * from './loaders';
+export * from './createStaticLoader';
